+++ tests/test_skool2html.py
@@ -54,6 +54,7 @@ class MockSkoolParser:
         self.entries = {}
         self.memory_map = []
         self.fields = {}
+        self.expands = ()
         mock_skool_parser = self
 
     def get_entry(self, address):